+++ internal/repo/user.go
@@ -7,20 +7,16 @@ import (
 	"log"
 	"strings"
 
-	"github.com/talenthandongsite/server-auth/internal/util"
-	"github.com/talenthandongsite/server-auth/pkg/enum/keychaintype"
-	"github.com/talenthandongsite/server-auth/pkg/jwt"
+	"github.com/talenthandongsite/server-auth/pkg/variable"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
-const DATABASE_NAME = "talent"
 const USER_COLLECTION_NAME = "user"
 
 type UserId struct{}
-type KeyType struct{}
 
 type User struct {
 	ID            string                  `json:"id,omitempty" bson:"_id,omitempty"`
@@ -37,43 +33,19 @@ type ActivityItem struct {
 	TimeStamp int64  `json:"timestamp,omitempty" bson:",omitempty"`
 }
 
-type KeyChainItem struct {
-	Content    string `json:"content,omitempty" bson:",omitempty"`
-	Secret     string `json:"secret,omitempty" bson:",omitempty"`
-	Expiration int64  `json:"expiration,omitempty" bson:",omitempty"`
-}
-
-type SignIn struct {
-	Username string `json:"username,omitempty" bson:",omitempty"`
-	Password string `json:"password,omitempty" bson:",omitempty"`
-}
-
-type DataItem struct {
-	Token string `json:"token,omitempty" bson:",omitempty"`
-	Exp   int64  `json:"exp,omitempty" bson:",omitempty"`
-}
-
-type SignInResponse struct {
-	Status string   `json:"status,omitempty" bson:",omitempty"`
-	Data   DataItem `json:"data,omitempty" bson:",omitempty"`
-}
-
-type JWTClaims struct {
-	ID            string `json:"id,omitempty" bson:"_id,omitempty"`
-	Username      string `json:"username,omitempty" bson:",omitempty"`
-	AccessControl string `json:"accessControl,omitempty" bson:",omitempty"`
-}
-
 type UserRepo struct {
 	Coll *mongo.Collection
-	Jwt  *jwt.Jwt
+	ctx  context.Context
 }
 
-func InitUserRepo(client *mongo.Client) *UserRepo {
-	database := client.Database(DATABASE_NAME)
+func InitUserRepo(ctx context.Context, client *mongo.Client) *UserRepo {
+	databaseName := variable.GetEnv(ctx, variable.DB_NAME)
+	database := client.Database(databaseName)
 	userCollection := database.Collection(USER_COLLECTION_NAME)
+
 	return &UserRepo{
 		Coll: userCollection,
+		ctx:  ctx,
 	}
 }
 
@@ -190,121 +162,4 @@ func (repo *UserRepo) Delete(ctx context.Context) (int, error) {
 		return 0, err
 	}
 	return int(deletedCount), nil
-}
-
-func (repo *UserRepo) ValidateUser(ctx context.Context, signin SignIn) (SignInResponse, error) {
-	log.Println("DEBUG : in repo ValidateUser")
-
-	var user User
-	err := repo.Coll.FindOne(
-		context.TODO(),
-		bson.M{"username": signin.Username},
-	).Decode(&user)
-
-	if err != nil {
-		// username에 해당되는 데이터가 없을 경우 false 반환
-		log.Println("DEBUG : in repo ValidateUser : Has no matching username")
-		if err == mongo.ErrNoDocuments {
-			return SignInResponse{
-				Status: "false",
-			}, err
-		}
-	}
-
-	token, expiration, err := repo.Jwt.ForgeToken(user.ID, user.Username, user.AccessControl)
-	if err != nil {
-		err := errors.New("token forge error")
-		log.Println("DEBUG : in repo ValidateUser : token forge error")
-		return SignInResponse{
-			Status: "false",
-		}, err
-	}
-
-	// username에 해당하는 데이터가 있고 비밀번호도 맞을 경우 리턴값 반환
-	return SignInResponse{
-		Status: "true",
-		Data: DataItem{
-			Token: "Bearer " + token,
-			Exp:   expiration.UnixMilli(),
-		},
-	}, nil
-}
-
-func (repo *UserRepo) UpsertKeychain(ctx context.Context, keychain *KeyChainItem) (bson.M, error) {
-
-	userId := fmt.Sprintf("%v", ctx.Value(UserId{}))
-	keyType := fmt.Sprintf("%v", ctx.Value(KeyType{}))
-
-	kct, err := keychaintype.Enum(keyType)
-	if err != nil {
-		return bson.M{}, err
-	}
-
-	objectId, err := primitive.ObjectIDFromHex(userId)
-	if err != nil {
-		return bson.M{}, err
-	}
-
-	log.Println("DEBUG : in repo upsert Keychain")
-
-	if kct == keychaintype.PASSWORD {
-		keychain.Content = util.HashSHA256(keychain.Content)
-		print(keychain.Content)
-	}
-
-	filter := bson.M{"_id": objectId}
-	update := bson.M{"$set": bson.M{"keychain." + keyType: keychain}}
-
-	upsert := true
-	after := options.After
-
-	opt := options.FindOneAndUpdateOptions{
-		ReturnDocument: &after,
-		Upsert:         &upsert,
-	}
-
-	result := repo.Coll.FindOneAndUpdate(ctx, filter, update, &opt)
-
-	if result.Err() != nil {
-		return nil, result.Err()
-	}
-
-	doc := bson.M{}
-	decodeErr := result.Decode(&doc)
-
-	return doc, decodeErr
-}
-
-func (repo *UserRepo) DeleteKeychain(ctx context.Context) (bson.M, error) {
-
-	log.Println("DEBUG : in repo Delete Keychain")
-
-	userId := fmt.Sprintf("%v", ctx.Value(UserId{}))
-	keyType := fmt.Sprintf("%v", ctx.Value(KeyType{}))
-
-	_, err := keychaintype.Enum(keyType)
-	if err != nil {
-		return nil, err
-	}
-
-	objectId, err := primitive.ObjectIDFromHex(userId)
-
-	if err != nil {
-		return nil, err
-	}
-
-	filter := bson.M{"_id": objectId}
-	update := bson.M{"$unset": bson.M{"keychain." + keyType: ""}}
-
-	after := options.Before
-
-	opt := options.FindOneAndUpdateOptions{
-		ReturnDocument: &after,
-	}
-	result := repo.Coll.FindOneAndUpdate(ctx, filter, update, &opt)
-
-	doc := bson.M{}
-	decodeErr := result.Decode(&doc)
-
-	return doc, decodeErr
 }
